Refactor the BeanFactory gadget code
