+++ modules/exploits/multi/http/log4shell_header_injection.rb
@@ -92,8 +92,8 @@ class MetasploitModule < Msf::Exploit::Remote
       OptString.new('TARGETURI', [ true, 'The URI to scan', '/']),
       OptString.new('HTTP_HEADER', [ false, 'The HTTP header to inject into' ]),
       OptEnum.new('JAVA_GADGET_CHAIN', [
-        true, 'The ysoserial payload to use for deserialization', 'CommonsBeanutils1',
-        Msf::Util::JavaDeserialization.ysoserial_payload_names
+        true, 'The Java gadget chain to use for deserialization', 'CommonsBeanutils1',
+        Msf::Exploit::JavaDeserialization.gadget_chains
       ], conditions: %w[TARGET != Automatic]),
       OptPort.new('HTTP_SRVPORT', [true, 'The HTTP server port', 8080], conditions: %w[TARGET == Automatic]),
       OptBool.new('LDAP_AUTH_BYPASS', [true, 'Ignore LDAP client authentication', true])
